Update BinarySearchForElementOccurringOnce.c
Given an array of n integers where every element appears twice except one element. Find that element. 

The idea is to use Binary Search(BS) to accomplish this. A simple XORing of the array will also do. 

For BS, if the target element is after a mid value , then start index of all other elements before it will be even and duplicate will be at odd.
 If the target element is before a mid value , then start index of all other elements before it will be odd and duplicate will be at even.