UniquePaths.java
A robot is located at the top-left corner of a m x n grid (marked 'Start' in the diagram below).

The robot can only move either down or right at any point in time. The robot is trying to reach the bottom-right corner of the grid (marked 'Finish' in the diagram below).

How many possible unique paths are there?
Imagine a 3x7 grid with 'Start' at (0,0) and 'Finish' at (2,6).